+++ docs/src/app/docs/components/layout.tsx
@@ -1,7 +1,7 @@
 import type { Metadata } from 'next';
 
 export const metadata: Metadata = {
-    metadataBase: new URL(`https://nordstar.dev/docs/components/`),
+    metadataBase: new URL(`https://nordcominc.github.io/nordstar/docs/components/`),
     title: {
         default: 'Components | Docs | Nordstar Component Library',
         template: `%s - Components | Docs | Nordstar Component Library`