+++ packages/core/nordstar/CHANGELOG.md
@@ -210,7 +210,7 @@
 
 - [`9dfeb3f`](https://github.com/NordcomInc/nordstar/commit/9dfeb3f0a233a67144d415d8d44d95d211807ea6) Thanks [@filiphsps](https://github.com/filiphsps)! - - Refactor components to use tailwind and `cva`.
 
-- [`9dfeb3f`](https://github.com/NordcomInc/nordstar/commit/9dfeb3f0a233a67144d415d8d44d95d211807ea6) Thanks [@filiphsps](https://github.com/filiphsps)! - - Update domain to `https://nordstar.dev/`
+- [`9dfeb3f`](https://github.com/NordcomInc/nordstar/commit/9dfeb3f0a233a67144d415d8d44d95d211807ea6) Thanks [@filiphsps](https://github.com/filiphsps)! - - Update domain to `https://nordcominc.github.io/nordstar/`
 
 - Updated dependencies [[`b2d7ad0`](https://github.com/NordcomInc/nordstar/commit/b2d7ad0e1b489a1229f7adf5752546b78dfef27d), [`8fbb855`](https://github.com/NordcomInc/nordstar/commit/8fbb855a1b85f7e6e2449b142702d95a002fa90a), [`0cb078a`](https://github.com/NordcomInc/nordstar/commit/0cb078a756ee53bb93606ed9237247d6715cc7ec), [`9dfeb3f`](https://github.com/NordcomInc/nordstar/commit/9dfeb3f0a233a67144d415d8d44d95d211807ea6), [`9dfeb3f`](https://github.com/NordcomInc/nordstar/commit/9dfeb3f0a233a67144d415d8d44d95d211807ea6), [`69a26c5`](https://github.com/NordcomInc/nordstar/commit/69a26c59e4d5492db2f39798ac1d628411fbb7d4), [`e586f33`](https://github.com/NordcomInc/nordstar/commit/e586f337007468d313af9741f9015d1c2a1f0aed)]:
     - @nordcom/nordstar-accented@0.0.63